Refactor editor components and add hooks
